Rename `verifier` to `check`
