+++ model-check/build.gradle.kts
@@ -1,0 +1,112 @@
+/*
+ * Copyright 2022, TeamDev. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+import io.spine.internal.gradle.publish.IncrementGuard
+
+buildscript {
+    apply(from = "$rootDir/version.gradle.kts")
+}
+
+plugins {
+    `maven-publish`
+    id("com.github.johnrengelman.shadow").version("7.1.2")
+}
+
+apply<IncrementGuard>()
+
+/** The publishing settings from the root project. */
+val spinePublishing = rootProject.the<io.spine.internal.gradle.publish.SpinePublishing>()
+
+val coreJavaVersion: String by extra
+val toolBaseVersion: String by extra
+val mcJavaVersion: String by extra
+val baseVersion: String by extra
+
+dependencies {
+    implementation(gradleApi())
+    implementation("io.spine:spine-server:$coreJavaVersion")
+    implementation("io.spine.tools:spine-plugin-base:$toolBaseVersion")
+
+    implementation(project(":model-assembler"))
+
+    testImplementation(gradleTestKit())
+    testImplementation("io.spine.tools:spine-testlib:$baseVersion")
+    testImplementation("io.spine.tools:spine-plugin-testlib:$toolBaseVersion")
+    testImplementation("io.spine.tools:spine-testutil-server:$coreJavaVersion")
+}
+
+tasks.test {
+    dependsOn("publishToMavenLocal",
+              ":model-assembler:publishToMavenLocal")
+}
+
+/**
+ * The ID of the far JAR artifact.
+ */
+val pArtifact = spinePublishing.artifactPrefix + "model-check-bundle"
+
+publishing {
+    val pGroup = project.group.toString()
+    val pVersion = project.version.toString()
+
+    publications {
+        create("fat-jar", MavenPublication::class) {
+            groupId = pGroup
+            artifactId = pArtifact
+            version = pVersion
+            artifact(tasks.shadowJar)
+        }
+    }
+}
+
+tasks.publish {
+    dependsOn(tasks.shadowJar)
+}
+
+tasks.shadowJar {
+    exclude(
+        /**
+         * Exclude Gradle types to reduce the size of the resulting JAR.
+         *
+         * Those required for the plugins are available at runtime anyway.
+         */
+        "org/gradle/**",
+
+        /**
+         * Remove all third-party plugin declarations as well.
+         *
+         * They should be loaded from their respective dependencies.
+         */
+        "META-INF/gradle-plugins/com**",
+        "META-INF/gradle-plugins/net**",
+        "META-INF/gradle-plugins/org**")
+
+    isZip64 = true  /* The archive has way too many items. So using the Zip64 mode. */
+    archiveClassifier.set("")  /** To prevent Gradle setting something like `osx-x86_64`. */
+    mergeServiceFiles("desc.ref")
+}
+
+
