+++ model-check/src/test/java/io/spine/model/check/given/InvalidRestoreAggregate.java
@@ -24,16 +24,27 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-pluginManagement {
-    repositories {
-        mavenCentral()
-        gradlePluginPortal()
+package io.spine.model.check.given;
+
+import io.spine.server.aggregate.Aggregate;
+import io.spine.server.command.Assign;
+import io.spine.test.model.verify.command.RestorePhoto;
+import io.spine.test.model.verify.event.PhotoRestored;
+import io.spine.test.model.verify.given.EditState;
+
+/**
+ * This aggregate declares a command-handling method that breaks the contract imposed by
+ * {@link Assign}, by having a {@code private} access modifier.
+ *
+ * <p>This should result in a warning.
+ */
+public class InvalidRestoreAggregate extends Aggregate<String, EditState, EditState.Builder> {
+
+    @SuppressWarnings("MethodMayBeStatic")
+    @Assign
+    private PhotoRestored handle(RestorePhoto restore){
+        return PhotoRestored.newBuilder()
+                .setTitle(restore.getTitle())
+                .build();
     }
 }
-
-rootProject.name = "spine-model-tools"
-
-include(
-    "model-assembler",
-    "model-check",
-)
