+++ model-check/src/test/java/io/spine/model/check/given/RenameProcMan.java
@@ -24,16 +24,33 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-pluginManagement {
-    repositories {
-        mavenCentral()
-        gradlePluginPortal()
+package io.spine.model.check.given;
+
+import io.spine.core.EventContext;
+import io.spine.server.aggregate.Apply;
+import io.spine.server.command.Assign;
+import io.spine.server.procman.ProcessManager;
+import io.spine.test.model.verify.command.ChangeTitle;
+import io.spine.test.model.verify.event.TitleChanged;
+import io.spine.test.model.verify.given.RenameState;
+
+public class RenameProcMan extends ProcessManager<String, RenameState, RenameState.Builder> {
+
+    protected RenameProcMan(String id) {
+        super(id);
+    }
+
+    @Assign
+    TitleChanged handle(ChangeTitle command) {
+        return TitleChanged
+                .newBuilder()
+                .setNewTitle(command.getNewTitle())
+                .build();
+    }
+
+    @Apply
+    private void on(TitleChanged event, EventContext context) {
+        builder().setEditor(context.actor()
+                                   .getValue());
     }
 }
-
-rootProject.name = "spine-model-tools"
-
-include(
-    "model-assembler",
-    "model-check",
-)
