+++ model-check/src/test/proto/spine/test/model/verify/commands.proto
