+++ model-check/src/test/proto/spine/test/model/verify/edit.proto
