+++ model-check/src/test/proto/spine/test/model/verify/events.proto
