+++ model-check/src/test/resources/model-check-test/src/main/java/io/spine/model/verify/ValidProcMan.java
@@ -24,16 +24,28 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-pluginManagement {
-    repositories {
-        mavenCentral()
-        gradlePluginPortal()
+package io.spine.model.check;
+
+import io.spine.server.command.Assign;
+import io.spine.server.procman.ProcessManager;
+
+import java.util.List;
+
+import static java.util.Collections.singletonList;
+
+/**
+ * A {@code ProcessManager} with a valid command-handling method.
+ */
+public class ValidProcMan extends ProcessManager<String, ValidProcess, ValidProcess.Builder> {
+
+    protected ValidProcMan(String id) {
+        super(id);
+    }
+
+    @Assign
+    List<VideoCallStarted> handle(StartVideoCall command) {
+        return singletonList(VideoCallStarted.newBuilder()
+                                             .setIp(command.getIp())
+                                             .build());
     }
 }
-
-rootProject.name = "spine-model-tools"
-
-include(
-    "model-assembler",
-    "model-check",
-)
