+++ model-check/src/test/resources/model-check-test/src/main/proto/spine/model/verify/call_entity.proto
