+++ model-check/src/test/resources/model-check-test/src/main/proto/spine/model/verify/commands.proto
