+++ model-check/src/test/resources/model-check-test/src/main/proto/spine/model/verify/events.proto
